worker: fix tests
