+++ packages/ws-worker/src/mock/runtime-engine.ts
@@ -98,6 +98,7 @@ async function createMock() {
           level: 'info',
           json: true,
           message: JSON.stringify(args),
+          name: 'JOB',
           time: Date.now(),
         });
       },
