+++ packages/ws-worker/test/events/step-start.test.ts
@@ -60,97 +60,3 @@ test('send a step:start event', async (t) => {
 
   await handleStepStart({ channel, state } as any, { jobId } as any);
 });
-
-test('step:start event should include versions', async (t) => {
-  const plan = {
-    id: 'run-1',
-    workflow: {
-      steps: [{ id: 'job-1', expression: '.' }],
-    },
-    options: {},
-  };
-  const input = 'abc';
-  const jobId = 'job-1';
-
-  const versions = {
-    node: process.version.substring(1),
-    engine: '1.0.0',
-    compiler: '1.0.0',
-    worker: pkg.version,
-  };
-
-  // Simulate an event that would be generated by the worker
-  const event: JobStartPayload = {
-    jobId,
-    workflowId: plan.id,
-    versions,
-  };
-
-  const state = createRunState(plan, input);
-  state.activeJob = jobId;
-  state.activeStep = 'b';
-
-  const channel = mockChannel({
-    [STEP_START]: (evt) => {
-      t.deepEqual(evt.versions, {
-        ...versions,
-        worker: pkg.version,
-      });
-      return true;
-    },
-    [RUN_LOG]: () => true,
-  });
-
-  await handleStepStart({ channel, state } as any, event);
-});
-
-test('also logs the version number', async (t) => {
-  let logEvent: any;
-  const plan = {
-    id: 'run-1',
-    workflow: {
-      steps: [{ id: 'job-1', expression: '.' }],
-    },
-    options: {},
-  };
-  const input = 'abc';
-  const jobId = 'job-1';
-
-  const versions = {
-    node: process.version.substring(1),
-    engine: '1.0.0',
-    compiler: '1.0.0',
-    worker: pkg.version,
-  };
-
-  // Simulate an event that would be generated by the worker
-  const event: JobStartPayload = {
-    jobId,
-    workflowId: plan.id,
-    versions,
-  };
-
-  const state = createRunState(plan, input);
-  state.activeJob = jobId;
-  state.activeStep = 'b';
-
-  const channel = mockChannel({
-    [STEP_START]: () => true,
-    [RUN_LOG]: (evt) => {
-      if (evt.source === 'VER') {
-        logEvent = evt;
-      }
-      return true;
-    },
-  });
-
-  await handleStepStart({ channel, state } as any, event);
-
-  t.truthy(logEvent);
-  t.is(logEvent.level, 'info');
-  const [message] = logEvent.message;
-  t.log(message);
-  // This just a light test of the string to make sure it's here
-  // It uses src/util/versions, which is tested elsewhere
-  t.regex(message, /(node\.js).+(worker).+(engine)/is);
-});
